+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Fix Accessibility problems for settings/data-import/{...-profiles} (Form elements must have labels) (UIDATIMP-457)
 * Fix Accessibility problems in ProfileLinker Component (settings/data-import/job-profiles) (UIDATIMP-434)
 * Cannot delete an import stuck in "Running" (UIDATIMP-738)
+* Fix matching by `id` for Holdings and Item (UIDATIMP-747)
 
 ## [3.0.1](https://github.com/folio-org/ui-data-import/tree/v3.0.1) (2020-10-27)